correctifs

+++ app/src/main/java/com/example/huski/gpsStruct.java
@@ -1,27 +1,27 @@
 package com.example.huski;
 
 public class gpsStruct {
-    private float lon;
-    private float lat;
+    private double lon;
+    private double lat;
 
-    public gpsStruct(float lon, float lat){
+    public gpsStruct(double lon, double lat){
         this.lon = lon;
         this.lat = lat;
     }
 
-    public float getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(float lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(float lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
